[Binary Encoding] Fix list encoding bug introduced in the refactor
Nested list decoding was broken when I tried to optimize too much
