+++ CCSDS_MAL_ENCODING_BINARY/src/main/java/esa/mo/mal/encoder/binary/fixed/FixedBinaryDecoder.java
@@ -22,7 +22,9 @@ package esa.mo.mal.encoder.binary.fixed;
 
 import esa.mo.mal.encoder.binary.base.BinaryTimeHandler;
 import java.math.BigInteger;
+import java.util.List;
 import org.ccsds.moims.mo.mal.MALException;
+import org.ccsds.moims.mo.mal.MALListDecoder;
 
 /**
  * Implements the MALDecoder interface for a fixed length binary encoding.
@@ -86,6 +88,12 @@ public class FixedBinaryDecoder extends esa.mo.mal.encoder.binary.base.BaseBinar
           final BinaryTimeHandler timeHandler)
   {
     super(src, timeHandler);
+  }
+
+  @Override
+  public MALListDecoder createListDecoder(final List list) throws MALException
+  {
+    return new FixedBinaryListDecoder(list, sourceBuffer, timeHandler);
   }
 
   /**
