+++ CCSDS_MAL_ENCODING_BINARY/src/main/java/esa/mo/mal/encoder/binary/variable/VariableBinaryDecoder.java
@@ -25,6 +25,7 @@ import java.io.InputStream;
 import java.math.BigInteger;
 import java.util.List;
 import org.ccsds.moims.mo.mal.MALException;
+import org.ccsds.moims.mo.mal.MALListDecoder;
 
 /**
  * Implements the MALDecoder interface for a fixed length binary encoding.
@@ -74,6 +75,12 @@ public class VariableBinaryDecoder extends esa.mo.mal.encoder.binary.base.BaseBi
   public VariableBinaryDecoder(final BufferHolder src, final BinaryTimeHandler timeHandler)
   {
     super(src, timeHandler);
+  }
+
+  @Override
+  public MALListDecoder createListDecoder(final List list) throws MALException
+  {
+    return new VariableBinaryListDecoder(list, sourceBuffer, timeHandler);
   }
 
   /**
